Modify logic for interactable key press Add boolean for triggering interactable code block
It is intended for a single key input to be used for attacking and
interacting with objects. This is currently the E key. When the player
enters a trigger field, isNearInteractable is set to true. On exiting
the trigger field, isNearInteractable is set to false.

Whilst isNearInteractable is true, pressing the interact key will
make the player interact instead of attack.