Add check for Interact tagged objects Add Interact in Tags list
OnTriggerEnter now stores the game object of the field being entered.
In addition, interactable objects are expected to be tagged with the
Interact tag. This tag is checked for when the user attempts to
interact whilst in a trigger field. The user will not interact if
the tag is not present.