Add switch case in Interactable
Added a switch case that checks the tag that the Interactable script
is attached to. Depending on the tag, a different method is called.
The different methods depends on the object being interacted with.

Different interaction actions will come with later features. For now,
the basic Interactable script is complete.