Add TimedGate class Add OpenTheGate method in TimedGate class Add OpenTheGate call in Interactable class
Created a new class, TimedGate, that will eventually allow gates
to be temporarily open. This assumes that the Switch object is a child
of the object that has TimedGate.

In Interactable class, the method PullSwitch searches in the parent
objects until one has the TimedGate component. Then it will call
TimedGate method OpenTheGate. At the moment, OpenTheGate only prints a
debug message.