Add method to disable collision on gate object
For TimedGate, a boolean checks whether the switch has been pulled yet.
If the switch has yet to be pulled, isSwitchPulled is set to true to
prevent multiple interactions with the switch.

For this prototype, the collision on the door is changed to disabled,
and the colour of the sprite is green.

The new method, CloseTheGate, is invoked after 5.0 seconds. This method
re-enables collision on the gate, changes the sprite colour to red, and
sets isSwitchedPulled to false. This allows the switch to be pulled
again.