Add tag check in Player Attack method Add method to check player damage threshold
Player method, Attack, now checks the object's tag before attempting
to find either the Enemy or BreakableBarricade component.

If it's a barricade, it will call a method to check that the player
can exceed the damage threshold value before damage is applied.
If the damage threshold is exceeded, the barricade takes damage.
Otherwise, the barricade loses no hp.

Barricade gameObject is destroyed when hp is 0 or less.