Add generalised upgrade tier method for button
The canvas menu buttons for selecting upgrades now have a script,
UpgradeTierButton. This script contains information on what trait the
button is supposed to affect (traitNum), and what tier it is unlocking
(traitTier). These are represented by integer values:
traitNum: 0 = jump; 1 = speed; 2 = strength.
traitTier:  0 = tier 0, 1 = tier 1, 2 = tier 2, 3 = tier 3.

There is a method which is called when the button is clicked which
sends traitNum and TraitTier to UpgradeMenu by calling a method in
UpgradeMenu, SetTargetButton.

SetTargetButton, accepts 2 integers, traitNum and traitTier, and
which are then used when UpgradeMenu calls Player method SetTier.