First commit adds double jump
Double jump check added to Player method Jump. New boolean value,
canJumpAgain, is set on the first jump from the ground. On the second
mid-air jump, canJumpAgain is set to false.

In addition, VelocityDecay has an additional check so that the player
doesn't drop as fast when the player has yet to use their second jump.