Add additional check in VelocityDecay method
Added a new variable, jumpTierFallReduction, that modifies the player's
fall speed. jumpTierFallReduction is set to a very small number if
the player has unlocked Tier-3 jump AND is holding the jump button down.
If the player does not have Tier-3 unlocked or has let go of the jump
button, the jumpTierFallReduction is set to 1f i.e. player will fall at
regular speed.