Add Charge Attack check in InteractOrAttack
InteractOrAttack method is now modified to check that whether the
attack key is held down. There is an additional check that increments
a timer variable. There are 2 more checks that determine whether a
normal or charged attack is performed depending on whether the timer
variable is large enough, the player has unlocked Tier-3 strength and
has let go of the attack key.

A charge attack doubles the damage of the current attack. The timer
variable is reset to 0f after performing an attack.

The Attack method now accepts an integer variable, damage. This is
because the damage value will change depending on whether a charge
attack or normal attack is performed.

In the Update method, the check for InteractOrAttack now checks for
either the continuous holding down of the attack key or when the attack
key is released.