Add Sprint check into Move and Update Modify player controls to apply velocity instead
In Update, there is now a check for whether the player is holding the
sprint key. If it is being held down AND the player has the Tier-2
upgrade, isSprinting is set to true. Otherwise, isSprinting is set to
false.

In Move method, there is now a conditional statement that runs when
isSprinting is true. An additional speed modifier is applied when the
player is sprinting. Otherwise, it will apply the normal speed bonuses.

Also, rather than using transform.Translate(moveAmount, 0, 0) to
move the player model,
rig.velocity = new Vector2 (moveAmount, rig.velocity.y) is being used
to make it consistent with the changes in the Jump method.