Use AddForce for movement Add isDashing check in movement
Changed how movement is applied to player from setting velocity
to using AddForce on the player object. This is because the velocity
set by general left/right movement would override the velocity set by
dashing.

The players movement velocity is capped depending on whether they are
walking, sprinting, or dashing.

There is a check in Update for checking if Dash is enabled by double
tapping the sprint button.