Add GameControl class
GameControl class is a singleton that will contain data related to the
player between scenes. The GameObject containing GameControl will not
be destoryed on loading the next scene and will persist within the
scene that is being left.

Then it will check for instances gameObjects with a GameControl class
If one is not present in the scene, the GameControl gameobject will be
added to the current scene.
If an different gameobject containing GameControl is found, the current
copy will destroy itself, and the most recent instance will be it.