Add unlockedTraits check in UpgradeTierButton
With the addition of the GameControl class storing player data,
a new check has been added in UpgradeTierButton in Start method.
The check looks at
GameControl.control.unlockedTraits[traitNum, TraitTier];
and see whether it is 1 (player has tier unlocked) or 0 (player has yet
to unlock tier). If 1, then the button is disabled. If 0, then the
button is enabled.