Remove y-velocity clamp in Move method
The diagonal jump's vertical component was shorter than a purely
vertical jump. The issue was in Move method, where the y-velocity
clamp was capping the y-velocity to be far less than what was specified
in Jump method when a horizontal input was detected.

For now, the y-velocity clamp in Move method has been commented out.
This makes the vertical component of the jump to be consistent when the
player is jumping from a standstill or whilst moving.