Add title menu scenes Add TitleMenu Class for title menu scene Fix issue in level3_scene1 for spawn
Added title menu scene. Currently, it starts the main game. Added
TitleMenu class for this scene.

Reordered Spawn gameobject location in DeadlyStuff object in
level3_scene1. Door class was spawning the user at the first GameObject
child in hierarchy, which was the Container gameobject. This would
position the player on the container, which is also where the loadscene
trigger is.