Fix for player spawns Modify text positions and canvas background
Player was spawning at the end of the level instead of the beginning.
This only occurred in the build and not when previewing the game in
Unity. The fix was to set the door IDs to be unique for each scene.

Intro background and text positions were modified to be more readable.